pyproject.toml
[tool.poetry]
name = "ape-simple-storage-v23"
version = "0.1.0"
description = ""
authors = ["PatrickAlphac <54278053+PatrickAlphaC@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ape_simple_storage_v23"}]

[tool.poetry.dependencies]
python = "^3.10,<3.11"
eth-ape = "^0.6.8"
titanoboa = {git = "https://github.com/vyperlang/titanoboa"}
ape-alchemy = "^0.6.1"
black = "^23.3.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
